LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-expected.txt



      new span two



    two
  two  two

FAIL The <selectedcontent> element should reflect the HTML contents of the selected <option>. assert_equals: The innerHTML of <selectedcontent> should be updated in response to a form reset. expected "\n      <span class=\"one\">span</span> one\n    " but got "\n      <span class=\"two\" data-foo=\"bar\">new span</span> two\n    "
FAIL When there are multiple <selectedcontent> elements, only the one in tree order should be kept up to date. assert_equals: Second selectedcontent should not be kept up to date. expected "" but got "two"
PASS <seletedcontent> behavior in disconnected <select>.
